refactor: 뷰 분리
