+++ src/main/kotlin/step3/Main.kt
@@ -1,7 +1,7 @@
 package step3
 
 import step3.controller.MainController
-import step3.view.ConsoleView
+import step3.view.console.ConsoleView
 
 fun main() {
     val controller = MainController(ConsoleView())
